move code, add more checks
